Create README.md
When compiling to .NET managed code, the compiler translates your source code into Microsoft intermediate language (MSIL), which is a CPU-independent set of instructions that can be efficiently converted to native code. Guardship .NET Protector is not a .NET code obfuscator. It transforms the revealed MSIL code of your executables into native unmanaged representation that keeps all .NET metadata unaffected. It grants the same layer of protection as native ASM and C++ code.